
\name{breast.example}
\alias{breast.example}


\title{
  Data for examples
}
\description{
 Data frame containing breast cancer data used in examples.
}
%\details{ }
\examples{

 data("breast.example", package="JPSurv")

 # Display first two rows
 breast.example[1:2,]
}
\keyword{data}
%\seealso{\code{\link{rebet}}}



