\name{prostate}
\docType{data}
\alias{prostate}
\title{Life Table for the Prostate Cancer}
\description{
	This data set is extracted from the SEER 9 Cancer Registries and it contains the life table of prostate cancer patients.
}
\usage{prostate}
\format{An R dataframe of the size 433x15}
\keyword{datasets}
