\name{dictionary.overview}
\alias{dictionary.overview}
\title{SEERStat Dictionary Overview}
\description{Reads the dictionary file and returns a parsed ovreview.}
\usage{dictionary.overview(DICfilename)}
\arguments{
\item{DICfilename}{File name/path for dictionary (.dic) file corresponding to SEERStat data file.}
}
\value{A list which contains the parsed information stored in the dictionary file.}

